Move ConnectionPool to store::pool module
This change moves the ConnectionPool enum to a store::pool module. As
such, it updates its generics to include the error type. This improves
is use with any future stores.

It is guarded by a new feature, "store", which is grandfathered-in as
stable and depended on by either "postgres" and "sqlite" features.
